requirements.txt
streamlit
pandas 
pillow 
plotly
psycopg2==2.9.10
psycopg2-binary==2.9.10
groq
